+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.0.32
 
 - The `bool optimizeForSize` parameter has been added to the `fastBuild` function. This allows to reduce the size of the generated parser code. At the moment, this refers to the generation of `Tag` parsers. Instead of fast and cumbersome (inlined) code, compact and slightly slow code (of calling a static extension method) will be generated
+- Minor improvements have been made to the source code since the addition of the `prefer_const_declarations` Dart SDK analyzer option
+- Minor improvements have been made to the `IdentifierExpression` parser builder
 
 ## 2.0.31